[Test] Add a shell script to install `py.test`
This patch adds a shell script to the `tests/` directory which will
install `py.test` on the developer's system, since that is the tool
used for Hypatia's test suite.  The rationale for putting this into a
separate script and not the existing installation scripts is that
pytest is *not* required for running Hypatia, only for developing it.
Therefore we do not want to install unnecessary dependencies for users
who are installing Hypatia with no intention of doing any serious
amount of development.

The installation script tries to install the Python 3 version of
pytest by testing for the existence of `pip3`.  The justification here
is that if `pip3` exists then it is very likely the developer prefers
to work on Hypatia using Python 3.  The script will fallback to the
Python 2.7 version of `pip` if `pip3` is not available.

The script relies on the use of redirecting output to `/dev/null` as
part of testing whether or not pytest installed successfully, which is
an aspect of the script that may break on Windows systems.  I do not
remember off the top of my head if Windows ports of shells like GNU
Bash handle `/dev/null` appropriately for that operating system.

Signed-off-by: Eric James Michael Ritz <ejmr@plutono.com>

GitHub-Issue: #42
See-Also: http://pytest.org/latest/