[Feature] Forcibly reinstall all packages when running test.sh
This patch modifies `test.sh` so that we run PIP with two additional
flags:

1. `--upgrade`
2. `--force-reinstall`

The second flag is the one most important to this change, but we still
need to use `--upgrade` to guarantee that `--force-reinstall` takes
effect.

The reason for this change is to strengthen the "clean slate" we want
every time we run `test.sh` for the test suite.  These changes
forcibly reinstall all required packages just in case...

1. Any required packages have updated since the last time we ran the
   test suite.

2. We have made any local changes to required packages for any
   experimental and/or testing reasons.  This is an uncommon situation
   but not unrealistic either.

Signed-off-by: Eric James Michael Ritz <ejmr@plutono.com>

Closes: #97