[Chore] Tell Git to ignore any local Python "eggs" directory
Depending on how we run the test suite we may end up with an `.eggs/`
directory in the project.  We do not want to commit any such files by
accident, and our project-wide Git ignore file already includes some
rules for ignoring eggs.  This patch adds the `.eggs/` directory to
that list to safeguard ourselves from accidentally committing anything
in that directory.

Signed-off-by: Eric James Michael Ritz <ejmr@plutono.com>

GitHub-Issue: #97